Rename `config` module to `ppdb_config`
This aligns with DM practices of module names aligning with the main
class that they define, which is in this case `PpdbConfig`. Camelcase
conventions are used following current guidelines.
