+++ python/lsst/dax/ppdb/__init__.py
@@ -19,7 +19,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .config import *
+from .ppdb_config import *
 from .ppdb import *
 from .replicator import *
 from .version import *  # Generated by sconsUtils
